Move code to other file
